backend/requirements_aws.txt
fastapi
uvicorn
chromadb
sentence-transformers
pydantic
Pillow
tqdm
boto3
